requirements.txt
jira
pyresume
pyparser
python-docx
docx2pdf
